--- src/components/repo-language/index.ts
@@ -1,3 +1,0 @@
-import RepoLanguage from "./RepoLanguage";
-
-export default RepoLanguage
