+++ src/components/repo-languages/index.ts
@@ -1,0 +1,3 @@
+import RepoLanguages from "./RepoLanguages";
+
+export default RepoLanguages
